@@ -4,7 +4,8 @@
     "node": "14.x"
   },
   "scripts": {
-    "build": "tsc -p web/tsconfig.json"
+    "build": "tsc -p web/tsconfig.json",
+    "lint": "eslint . --ext .ts"
   },
   "dependencies": {
     "chrome-aws-lambda": "7.0.0",
@@ -16,6 +17,10 @@
     "@types/marked": "4.0.1",
     "@types/puppeteer": "5.4.4",
     "@types/puppeteer-core": "5.4.0",
-    "typescript": "4.1.5"
+    "@types/twemoji": "13.1.2",
+    "@typescript-eslint/eslint-plugin": "5.30.7",
+    "@typescript-eslint/parser": "5.30.7",
+    "eslint": "8.20.0",
+    "typescript": "4.7.4"
   }
 }
